Moved 6 funcs to `pronouns.py`: get_pronouns, get_new_pronouns, determine_plural_pronouns, get_default_pronouns, get_custom_pronouns, add_custom_pronouns
