+++ tests/test_lang.py
@@ -10,10 +10,9 @@ os.environ["SDL_AUDIODRIVER"] = "dummy"
 
 from scripts.cat.cats import Cat
 from scripts.game_structure.localization import (
-    get_new_pronouns,
-    determine_plural_pronouns,
     set_lang_config_directory,
 )
+from scripts.cat.pronouns import get_new_pronouns, determine_plural_pronouns
 from scripts.utility import event_text_adjust
